added Requestor classes changed privacy api updated osgi refs updated upm test

+++ privacy-trust/java/privacy-protection/api/src/main/java/org/societies/privacytrust/privacyprotection/api/model/privacypreference/IDObfAction.java
@@ -35,6 +35,6 @@ package org.societies.privacytrust.privacyprotection.api.model.privacypreference
  */
 public interface IDObfAction {
 
-	public int getObfuscationLevel();
+	public double getObfuscationLevel();
 
 }
